README: describe what './new' script documents
The knowledge required to implement blktests test cases are documented
in './new' script which generates test case script template. This is
handy when we implement new test cases. However, it is difficult to
refer the documentation when we do not implement new test cases.

To help to refer the documentation, add a pointer to it and describe
what it documents.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>