check: enable bash error exit option
Currently, blktests test cases keep on running when bash errors happen,
then the errors are tend to be overlooked. Enable bash error exit option
with 'set -e' command to catch the bash errors in each test case. Also
enable inherit_errexit option to propagate it to sub-shells.

The error exit option does not work for the bash context in "if"
condition or in sub-shell (). Modify _check() and _call_test() so that
errexit work in test case contexts. Enable the errexit only in test()
and test_device() so that errexit gets disabled in requires() and its
friend functions because those functions are designed to cause error for
requirement checks.

Of note is that this change causes failures and weird behavior of many
test cases. Fixes in the affected test cases should be made together.

This change cause some clean-up actions skipped at error exits from test
cases. Some more improvement is desired.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>