nvme/032: fix device requirement check
The test case overrides global variable nvme_trtype. Depending on the
blktests framework implementation, this affects following test cases.
Remove the global variable change.

Instead, require the default transport type 'loop' so that this case is
run for the default set up, and not run for other fabric transports.

Also add a missing requirement to ensure that the test target device is
a PCI device.

(Is it really good that global variable changes by each test to affect
 following test cases?)

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>