common/xfs: check mount status before umount
The helper function _xfs_mkfs_and_mount() unmounts the mount target path
before mount operation to ensure that the path is not used for previous
mount operation. This unmount operation often fails, since the path is
not yet mounted for most cases. This failure causes script stop when
bash errexit option is enabled. To avoid the failure, ignore the umount
command exit status.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>