common/scsi_debug: check the write permission of sysfs attributes
scsi_debug has two sets of sysfs attributes: one is placed under the
module path (/sys/module/scsi_debug/parameters), and the other is placed
under the pseudo driver path (/sys/bus/pseudo/drivers/scsi_debug). Each
scsi_debug parameter is placed in one of the two sets, or both.
The function _scsi_debug_key_path() checks the existence of the
attribute file of the given parameter in the two sets, and return the
path to the attribute file available. However, it does not check write
permission of the files. _scsi_debug_key_path() may return the file not
writable, even when the file in the other set is writable.

To allow writes to such parameters, check the write permission of the
attribute files. If one of them is writable, return the writable file.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>