nvme: fix fc port clean up failure
Commit "nvme: sanitize transport parameter in _create_nvmet_port()"
modified the wwnn and the wwpn names to be calculated based on port ID.
However, this port dependent wwnn and wwpn calculation was not reflected
to the clean up steps. This made the test case nvme/031 fail. To avoid
the failure, add clean up steps based on the port ID.

Reviewed-by: Daniel Wagner <dwagner@suse.de>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>