common/nvme, nvme/003: specify "--port none" to _nvme_connect_subsys()
Commit "nvme: pass in port parameter to _nvme_connect_subsys()"
introduced the --port argument to _nvme_connect_subsys(). If the option
is not specified, _nvme_connect_subsys() automatically finds the port
for the specified subsysnqn. However, this does not work for the test
case nvme/003, since the test case connects to the subsys for discovery,
then, the port of the created subsysnqn is not to be used. This causes
the failure of the test case.

To avoid the failure, modify _nvme_connect_subsys() to take the special
value "none" for the --port option. When "--port none" is specified,
_nvme_connect_subsys() skips the steps to obtain port dependent
parameters and just use default parameters.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>