block/032: make error messages clearer if mkfs or mount fails
If block/032 fails at mkfs we want to know why, so propagate
error messages. While at it, enhance the test to also propagate
the error return from mount and remove the odd sleep for a udevadm
settle as that's the only thing I can think of we need to wait for
here.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
[Shin'ichiro: enriched _xfs_mkfs_and_mount failure handling]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>