common/xfs: add _test_dev_suits_xfs() to verify logical block size will work
mkfs.xfs will use the sector size exposed by the device, if this
is larger than 32k this will fail as the largest sector size on XFS
is 32k. Provide a sanity check to ensure we skip creating a filesystem
if the sector size is larger than what XFS supports.

Suggested-by: Shinichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
[Shin'ichiro: drop changes for block/032 and nvme/012]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>